Refactored auth into separate Login component
